Add inference scenario
