+++ benchmarking/common/helpers.js
@@ -1,0 +1,6 @@
+function readTestData(name) {
+  return {
+    rest: JSON.parse(open(`../data/${name}/rest-requests.json`)),
+    // grpc: JSON.parse(open(`../data/${name}/grpc-requests.pb`)),
+  };
+}
